Refactor: Split chart components into smaller files
This commit refactors the chart components in src/components/ChartComponents.tsx into smaller, more manageable files. This improves code organization and maintainability. Unused imports and files are also removed.
