+++ src/components/charts/index.ts
@@ -1,0 +1,8 @@
+
+// Export all chart components from a single file for easier imports
+export { ChartContainer } from './ChartContainer';
+export { SimpleBarChart } from './SimpleBarChart';
+export { SimplePieChart } from './SimplePieChart';
+export { SimpleLineChart } from './SimpleLineChart';
+export { SimpleRadarChart } from './SimpleRadarChart';
+export { getRandomData } from './utils';
